requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cloudinary==1.36.0
Django==4.2.7
django-autoslug==1.9.9
django-debug-toolbar==4.2.0
django-jazzmin==2.6.0
Pillow==10.1.0
psycopg2==2.9.9
python-decouple==3.8
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
